# Todolistenverwaltung_openAPI.yaml
openapi: 3.0.1
info:
  title: Todo-Listen-Verwaltung
  description: ""
  contact:
    email: wichmann@bbs-os-brinkstr.de
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.1.0
servers:
- url: http://127.0.0.1:5000/
tags:
- name: Listenverwaltung
paths:
  /list/{list_id}:
    get:
      tags:
      - Listenverwaltung
      summary: Get all items from list
      operationId: getList
      parameters:
      - name: list_id
        in: path
        description: list id
        required: true
        schema:
          type: string
      responses:
        200:
          description: List item returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TodoEntry'
        404:
          description: Invalid list id
          content: {}
    delete:
      tags:
      - Listenverwaltung
      summary: Delete a list
      operationId: deleteList
      parameters:
      - name: list_id
        in: path
        description: list id
        required: true
        schema:
          type: string
      responses:
        200:
          description: List was deleted
          content: {}
        404:
          description: Invalid list id
          content: {}
  /list:
    post:
      tags:
      - Listenverwaltung
      summary: Add new list
      operationId: addList
      requestBody:
        description: list object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoList'
        required: true
      responses:
        200:
          description: List added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoList'
        500:
          description: List could not be added
          content: {}
      x-codegen-request-body-name: body
  /list/{list_id}/item:
    post:
      tags:
      - Listenelementverwaltung
      summary: Add new list item
      operationId: addListItem
      parameters:
      - name: list_id
        in: path
        description: list id
        required: true
        schema:
          type: string
      requestBody:
        description: list item object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewTodoEntry'
      responses:
        200:
          description: List item returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoEntry'
        404:
          description: Invalid list id
          content: {}
        500:
          description: Item could not be added
          content: {}
      x-codegen-request-body-name: body
  /list/{list_id}/item/{item_id}:
    parameters:
    - name: list_id
      in: path
      description: list id
      required: true
      schema:
        type: string
    - name: item_id
      in: path
      description: item id
      required: true
      schema:
        type: string
    patch:
      tags:
      - Listenelementverwaltung
      summary: Edit list item
      operationId: editListItem
      requestBody:
        description: list item object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewTodoEntry'
      responses:
        200:
          description: List item returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoEntry'
        400:
          description: Invalid parameters
          content: {}
        404:
          description: Invalid list id
          content: {}
        500:
          description: Item could not be added
          content: {}
    delete:
      tags:
      - Listenelementverwaltung
      summary: Delete a list item
      operationId: deleteListItem
      responses:
        200:
          description: List item was deleted
          content: {}
        404:
          description: Invalid list item id
          content: {}
        500:
          description: Item could not be deleted
          content: {}
  /lists:
    get:
      tags:
      - Listenverwaltung
      summary: Get all lists
      operationId: getLists
      responses:
        200:
          description: Lists returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TodoList'
components:
  schemas:
    TodoList:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    NewTodoEntry:
      type: object
      properties:
        name:
          type: string
        beschreibung:
          type: string
    TodoEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        user_id:
          type: string
          format: uuid
        list_id:
          type: string
          format: uuid